Refactor edit roles preparation functions into useRolesProposalFunctions
