+++ src/pages/daos/[daoAddress]/roles/edit/index.tsx
@@ -4,21 +4,12 @@ import { FieldArray, Form, Formik } from 'formik';
 import { useCallback, useMemo, useState } from 'react';
 import { useTranslation } from 'react-i18next';
 import { toast } from 'react-toastify';
-import { Address, encodeFunctionData, zeroAddress } from 'viem';
-import DecentHatsAbi from '../../../../../assets/abi/DecentHats';
-import GnosisSafeL2 from '../../../../../assets/abi/GnosisSafeL2';
+import { zeroAddress } from 'viem';
 import { RoleCard } from '../../../../../components/pages/Roles/RoleCard';
 import { RolesTable } from '../../../../../components/pages/Roles/RolesTable';
 import RoleFormCreateProposal from '../../../../../components/pages/Roles/forms/RoleFormCreateProposal';
 import RoleForm from '../../../../../components/pages/Roles/forms/RoleFormTabs';
-import {
-  RoleFormValues,
-  DEFAULT_ROLE_HAT,
-  HatStruct,
-  RoleValue,
-  EditBadgeStatus,
-  HatStructWithId,
-} from '../../../../../components/pages/Roles/types';
+import { RoleFormValues, DEFAULT_ROLE_HAT } from '../../../../../components/pages/Roles/types';
 import { Card } from '../../../../../components/ui/cards/Card';
 import { BarLoader } from '../../../../../components/ui/loaders/BarLoader';
 import PageHeader from '../../../../../components/ui/page/Header/PageHeader';
@@ -26,10 +17,11 @@ import { useHeaderHeight } from '../../../../../constants/common';
 import { DAO_ROUTES } from '../../../../../constants/routes';
 import useSubmitProposal from '../../../../../hooks/DAO/proposal/useSubmitProposal';
 import { useRolesSchema } from '../../../../../hooks/schemas/roles/useRolesSchema';
+import { useRolesProposalFunctions } from '../../../../../hooks/utils/useRolesProposalFunctions';
 import { useFractal } from '../../../../../providers/App/AppProvider';
 import { useNetworkConfig } from '../../../../../providers/NetworkConfig/NetworkConfigProvider';
 import { useRolesState } from '../../../../../state/useRolesState';
-import { CreateProposalMetadata, ProposalExecuteData } from '../../../../../types';
+import { ProposalExecuteData } from '../../../../../types';
 
 function RolesEdit() {
   const { t } = useTranslation(['roles', 'navigation', 'breadcrumbs', 'dashboard']);
@@ -44,6 +36,9 @@ function RolesEdit() {
   const { hatsTree, hatsTreeId } = useRolesState();
 
   const { submitProposal } = useSubmitProposal();
+
+  const { parsedEditedHats, prepareCreateTopHatProposal, prepareEditHatsProposal } =
+    useRolesProposalFunctions();
 
   const hats = useMemo(() => {
     // @todo get hats from hatsTree from state
@@ -76,191 +71,6 @@ function RolesEdit() {
     setHatIndex(_hatIndex);
   };
 
-  const decentHatsAddress = '0x88e72194d93bf417310b197275d972cf78406163'; // @todo: sepolia only. Move to, and read from, network config
-
-  const prepareCreateTopHatProposal = useCallback(
-    async (proposalMetadata: CreateProposalMetadata, addedHats: HatStruct[]) => {
-      if (!safe) return;
-
-      const enableModuleData = encodeFunctionData({
-        abi: GnosisSafeL2,
-        functionName: 'enableModule',
-        args: [decentHatsAddress],
-      });
-
-      const disableModuleData = encodeFunctionData({
-        abi: GnosisSafeL2,
-        functionName: 'disableModule',
-        args: [decentHatsAddress, decentHatsAddress], // @todo: Figure out prevModule arg. Need to retrieve from safe.
-      });
-
-      const addressZero = zeroAddress as Address;
-      const adminHat: HatStruct = {
-        eligibility: addressZero,
-        toggle: addressZero,
-        maxSupply: 1,
-        details: JSON.stringify({
-          name: 'Admin',
-          description: '',
-        }),
-        imageURI: '',
-        isMutable: true,
-        wearer: addressZero,
-      };
-
-      const createAndDeclareTreeData = encodeFunctionData({
-        abi: DecentHatsAbi,
-        functionName: 'createAndDeclareTree',
-        args: [
-          JSON.stringify({
-            name: 'Top Hat',
-            description: 'top hat',
-          }),
-          '',
-          adminHat,
-          addedHats,
-        ],
-      });
-
-      return {
-        targets: [safe.address, decentHatsAddress, safe.address] as Address[],
-        calldatas: [enableModuleData, createAndDeclareTreeData, disableModuleData],
-        metaData: proposalMetadata,
-        values: [0n, 0n, 0n],
-      };
-    },
-    [safe],
-  );
-
-  const prepareAddHatsArgs = useCallback((addedHats: HatStruct[]) => {
-    const admins: bigint[] = [];
-    const details: string[] = [];
-    const maxSupplies: number[] = [];
-    const eligibilityModules: Address[] = [];
-    const toggleModules: Address[] = [];
-    const mutables: boolean[] = [];
-    const imageURIs: string[] = [];
-
-    addedHats.forEach(hat => {
-      admins.push(0n);
-      details.push(hat.details);
-      maxSupplies.push(hat.maxSupply);
-      eligibilityModules.push(hat.eligibility);
-      toggleModules.push(hat.toggle);
-      mutables.push(hat.isMutable);
-      imageURIs.push(hat.imageURI);
-    });
-
-    return [admins, details, maxSupplies, eligibilityModules, toggleModules, mutables, imageURIs];
-  }, []);
-
-  const HatsAbi = 'HatsAbi' as any;
-  const hatsContractAddress = 'hatsAddress' as Address;
-
-  const prepareEditHatsProposal = useCallback(
-    async (
-      treeId: number,
-      proposalMetadata: CreateProposalMetadata,
-      edits: {
-        addedHats: HatStruct[];
-        removedHatIds: number[];
-        updatedHats: HatStructWithId[];
-      },
-    ) => {
-      if (!safe) return;
-
-      const { addedHats, removedHatIds, updatedHats } = edits;
-
-      // @todo: assert none of edited hats are top hat or admin hat
-      if (removedHatIds.includes(treeId) || updatedHats.some(hat => hat.id === treeId)) {
-        throw new Error('Cannot edit top hat');
-      }
-
-      const addHatsTx = encodeFunctionData({
-        abi: 'HatsAbi' as any,
-        functionName: 'batchCreateHats',
-        args: prepareAddHatsArgs(addedHats),
-      });
-
-      const removeHatTxs = removedHatIds.map(hatId =>
-        encodeFunctionData({
-          abi: HatsAbi,
-          functionName: 'setHatStatus',
-          args: [hatId, false],
-        }),
-      );
-
-      return {
-        targets: [hatsContractAddress, ...removeHatTxs.map(() => hatsContractAddress)],
-        calldatas: [addHatsTx, ...removeHatTxs],
-        metaData: proposalMetadata,
-        values: [0n, ...removeHatTxs.map(() => 0n)],
-      };
-    },
-    [prepareAddHatsArgs, safe],
-  );
-
-  const parsedEditedHats = (editedHats: RoleValue[]) => {
-    const addedHats: HatStruct[] = [];
-    const removedHatIds: number[] = [];
-    const updatedHats: HatStructWithId[] = [];
-
-    editedHats.forEach(hat => {
-      const { roleName, member, roleDescription, editedRole, id } = hat;
-      if (editedRole) {
-        switch (editedRole.status) {
-          case EditBadgeStatus.New:
-            addedHats.push({
-              eligibility: zeroAddress,
-              toggle: zeroAddress,
-              maxSupply: 1,
-              details: JSON.stringify({
-                name: roleName,
-                description: roleDescription,
-              }),
-              imageURI: '',
-              isMutable: true,
-              wearer: member as Address,
-            });
-            break;
-
-          case EditBadgeStatus.Removed:
-            removedHatIds.push(id);
-            break;
-
-          case EditBadgeStatus.Updated:
-            editedRole.fieldNames.forEach(fieldName => {
-              switch (fieldName) {
-                case 'roleName':
-                  break;
-                case 'roleDescription':
-                  break;
-                case 'member':
-                  break;
-              }
-            });
-
-            updatedHats.push({
-              id,
-              eligibility: zeroAddress,
-              toggle: zeroAddress,
-              maxSupply: 1,
-              details: JSON.stringify({
-                name: roleName,
-                description: roleDescription,
-              }),
-              imageURI: '',
-              isMutable: true,
-              wearer: member as Address,
-            });
-            break;
-        }
-      }
-    });
-
-    return { addedHats, removedHatIds, updatedHats };
-  };
-
   const createRolesEditProposal = useCallback(
     async (values: RoleFormValues) => {
       try {
@@ -270,9 +80,9 @@ function RolesEdit() {
         let proposalData: ProposalExecuteData | undefined;
 
         const { addedHats, removedHatIds, updatedHats } = parsedEditedHats(modifiedHats);
-        
+
         if (hatsTreeId === null || hatsTreeId === undefined) {
-          // This safe has no top hat, so we prepare a proposal to create one. This will also create an admin hat, 
+          // This safe has no top hat, so we prepare a proposal to create one. This will also create an admin hat,
           // along with any other hats that are added.
           proposalData = await prepareCreateTopHatProposal(values.proposalMetadata, addedHats);
         } else {
@@ -303,6 +113,7 @@ function RolesEdit() {
       }
     },
     [
+      parsedEditedHats,
       hatsTreeId,
       prepareCreateTopHatProposal,
       prepareEditHatsProposal,
